+++ SharePointFramework/ProjectWebParts/src/data/SPDataAdapter/index.ts
@@ -61,8 +61,8 @@ export default new (class SPDataAdapter extends SPDataAdapterBase<ISPDataAdapter
       const [fields, siteUsers] = await Promise.all([
         templateParameters.ProjectContentTypeId
           ? this.entityService
-              .usingParams({ contentTypeId: templateParameters.ProjectContentTypeId })
-              .getEntityFields()
+            .usingParams({ contentTypeId: templateParameters.ProjectContentTypeId })
+            .getEntityFields()
           : this.entityService.getEntityFields(),
         this.sp.web.siteUsers
           .select('Id', 'Email', 'LoginName')
@@ -105,17 +105,26 @@ export default new (class SPDataAdapter extends SPDataAdapterBase<ISPDataAdapter
             break
           case 'User':
             {
-              const [siteUser] = siteUsers.filter(
+              const [_user] = siteUsers.filter(
                 (u) => u.Id === fieldValues[`${fld.InternalName}Id`]
               )
-              const user = siteUser
-                ? await this.entityService.web.ensureUser(siteUser.LoginName)
-                : null
+              const user = _user ? await this.entityService.web.ensureUser(_user.LoginName) : null
               properties[`${fld.InternalName}Id`] = user ? user.data.Id : null
             }
             break
+          case 'UserMulti': {
+            const userIds = fieldValues[`${fld.InternalName}Id`] || []
+            const users = siteUsers.filter(
+              (u) => userIds.indexOf(u.Id) !== -1
+            )
+            const ensured = await Promise.all(
+              users.map(({ LoginName }) => this.entityService.web.ensureUser(LoginName))
+            )
+            properties[`${fld.InternalName}Id`] = { results: ensured.map(({ data }) => data.Id) }
+          }
+            break
           case 'DateTime':
-            {
+            { 
               properties[fld.InternalName] = fldValue ? new Date(fldValue) : null
             }
             break