+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Every change is marked with Pull Request ID.
 - Support adding AD groups to get porfolio insights from SP group #332, #352
 - Change to latest statusreport when creating a new statusreport #343
 - Issue were user couldn't exit the portfolio filter pane #353
+- Issues with custom project fields #378
 
 ## 1.2.3 - 2020-10-07